requirements.txt
accelerate==1.4.0
ace_tools==0.0
acres==0.2.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.45.2
anyio==4.8.0
appdirs==1.4.4
asttokens==3.0.0
attrs==25.1.0
blinker==1.9.0
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
colorama==0.4.6
colpali_engine @ git+https://github.com/illuin-tech/colpali@4241ca9ece4149342beac5d7e61fa2f378b16a36
comm==0.2.2
ConfigArgParse==1.7
configobj==5.0.9
configparser==7.1.0
contourpy==1.3.1
cycler==0.12.1
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dataclasses-json==0.6.7
datasets==3.2.0
debugpy==1.8.12
decorator==5.1.1
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
etelemetry==0.3.1
executing==2.2.0
faiss-cpu==1.9.0.post1
fastjsonschema==2.21.1
filelock==3.17.0
Flask==3.0.3
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.9.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.66.0
GPUtil==1.4.0
greenlet==3.1.1
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.27.1
idna==3.10
importlib_metadata==8.6.1
ipykernel==6.29.5
ipynb==0.5.1
ipython==8.31.0
ipywidgets==8.1.5
isodate==0.6.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
langchain==0.3.17
langchain-community==0.3.16
langchain-core==0.3.34
langchain-openai==0.3.4
langchain-text-splitters==0.3.5
langsmith==0.3.4
looseversion==1.3.0
lxml==5.3.0
MarkupSafe==3.0.2
marshmallow==3.26.0
matplotlib==3.10.0
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
narwhals==1.26.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.9.2
numpy>=1.25.0,<3.0.0
openai==1.60.1
opencv-python==4.11.0.86
orjson==3.10.15
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathlib==1.0.1
pdf2image==1.17.0
peft==0.14.0
pillow==11.1.0
platformdirs==4.3.6
plotly==6.0.0
prompt_toolkit==3.0.50
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.3
prov==2.0.1
psutil==6.1.1
pure_eval==0.2.3
puremagic==1.28
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydot==3.0.4
Pygments==2.19.1
PyMuPDF==1.25.3
pyparsing==3.2.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
pyxnat==1.6.2
PyYAML==6.0.2
pyzmq==26.2.0
ragas==0.2.13
rdflib==6.3.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
retrying==1.3.4
rpds-py==0.22.3
rsa==4.9
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
sentence-transformers==3.4.1
sentencepiece==0.2.0
setuptools==75.8.0
simplejson==3.19.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
stack-data==0.6.3
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.21.0
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
traits==7.0.2
transformers==4.49.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
wcwidth==0.2.13
Werkzeug==3.0.6
wheel==0.45.1
widgetsnbextension==4.0.13
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
# Windows only dependency
pywin32==308; sys_platform == 'win32'
fpdf2==2.0.0
unstructured==0.1.0
pdfminer.six==20221105
pi_heif==0.1.0
unstructured_inference==0.1.0
unstructured_pytesseract==0.1.0